requirements.txt
streamlit
pytesseract
pillow
sentence-transformers
supabase
numpy